refactor(dynamic-forms): unify derivation + property-derivation collectors
Replace the two independent tree traversals with a single
`collectAllDerivations` that walks the field tree once and emits both the
value-derivation and property-derivation slices. The two collectors
previously had the same `traverseFieldsWithContext` setup, the same
array/group ancestry tracking, and the same `buildEffectiveFieldKey` /
`buildPropertyOverrideKey` shape (verified by inspection — both produce
`${arrayPath}.$.${groupPath}.${fieldKey}` for the same input). The
duplication was load-bearing only as a historical accident.

Changes:
  * New `core/derivation/collection-context.ts`: extracts the shared
    `CollectionContext` shape, `buildEffectiveFieldKey`,
    `buildNearestGroupPath`, `resolveTokenDependencies`, the `$self` /
    `$group` tokens, and the standard `traverseFieldsWithContext`
    onArrayChild/onGroupChild handlers.
  * `core/derivation/derivation-collector.ts` becomes the canonical home
    for ALL derivation collection — value AND property. Exposes:
      - `collectDerivations(fields)` — value-only (existing API, now a
        thin slice of the unified walk).
      - `collectAllDerivations(fields, logger, tracker)` — single
        traversal emitting both slices. The new preferred entry point.
      - Per-field handlers (`collectValueEntriesFromField`,
        `collectPropertyEntriesFromField`) + entry builders
        (`createShorthandEntry`, `createValueLogicEntry`,
        `createPropertyLogicEntry`).
  * `core/property-derivation/property-derivation-collector.ts` is now
    21 lines: a thin wrapper that calls `collectAllDerivations(...).property`.
  * No semantic change to either collector's emitted entries — verified by
    the existing 30+ test cases in `property-derivation-collector.spec.ts`
    and the value-collector specs (all 3062 unit tests pass).

Numbers:
  * dynamic-forms fesm2022 bundle: 1,059,467 → 1,058,139 bytes (-1,328 / -0.13%)
  * Source LOC across collectors: 358 → 158 (-200) for the property
    collector + duplicate-traversal removal; offset by ~250 LOC in the
    new collection-context module for a net ~+50 in source. The bundle
    win comes from the bundler dedupe-ing the now-shared traversal +
    helper functions (previously two copies of the same code).
  * 3062 / 3062 unit tests pass

Follow-up — the applicator merge (Phase C: dispatch by `targetProperty`)
remains. The two applicators write to fundamentally different sinks
(FieldTree vs PropertyOverrideStore) so the unification will only share
the shared evaluation step (`evaluateCondition` + `computeValueFromEntry`,
already in `core/expressions/`), not the write step. Tracked separately.
