+++ packages/dynamic-forms/src/lib/core/derivation/collection-context.ts
@@ -1,0 +1,137 @@
+import { DynamicFormError } from '../../errors/dynamic-form-error';
+import { Logger } from '../../providers/features/logger/logger.interface';
+import type { WarningTracker } from '../../utils/warning-tracker';
+
+/**
+ * Tokens recognised by the dependency resolver. Exported so the few callers
+ * that build `dependsOn` arrays programmatically (e.g. tests) don't have to
+ * inline the string literal.
+ *
+ * @public
+ */
+export const SELF_DEPENDENCY_TOKEN = '$self';
+
+/**
+ * Token resolving to the absolute path of the field's nearest parent
+ * container (group or array).
+ *
+ * @public
+ */
+export const GROUP_DEPENDENCY_TOKEN = '$group';
+
+/**
+ * Context threaded through {@link traverseFieldsWithContext} during derivation
+ * collection. Tracks array + group ancestry so absolute field paths and
+ * `$self` / `$group` tokens can be resolved without re-walking the tree.
+ *
+ * `logger` + `tracker` are present only because the property-derivation
+ * variant historically needed them inline; both fields are optional so the
+ * value-only call site (which doesn't need either) can pass `{}`.
+ *
+ * @internal
+ */
+export interface CollectionContext {
+  /** Path to the enclosing array (e.g. `'items'` or `'orders.0.lineItems'`). */
+  arrayPath?: string;
+  /**
+   * Dot-joined ancestor group keys scoped to the current array item (or
+   * top-level when not inside an array). Reset at array boundaries — the
+   * array placeholder path takes over, and group ancestors OUTSIDE an
+   * array don't contribute to descendants' entry keys.
+   */
+  groupPath?: string;
+  /** Available for diagnostics by handlers that emit warnings. */
+  logger?: Logger;
+  /** Available for de-duplicated warnings by handlers that need it. */
+  tracker?: WarningTracker;
+}
+
+/**
+ * Builds the absolute field path for an entry whose enclosing context is
+ * tracked by {@link CollectionContext}.
+ *
+ * Shape ladder:
+ * - inside array AND groups: `${arrayPath}.$.${groupPath}.${fieldKey}`
+ * - inside array only:       `${arrayPath}.$.${fieldKey}`
+ * - inside groups only:      `${groupPath}.${fieldKey}`
+ * - root:                    `${fieldKey}`
+ *
+ * Both value and property collectors used to compute this independently; they
+ * produced identical strings, so the logic now lives here.
+ *
+ * @internal
+ */
+export function buildEffectiveFieldKey(fieldKey: string, context: CollectionContext): string {
+  if (context.arrayPath && context.groupPath) {
+    return `${context.arrayPath}.$.${context.groupPath}.${fieldKey}`;
+  }
+  if (context.arrayPath) {
+    return `${context.arrayPath}.$.${fieldKey}`;
+  }
+  if (context.groupPath) {
+    return `${context.groupPath}.${fieldKey}`;
+  }
+  return fieldKey;
+}
+
+/**
+ * Resolves the absolute path of the field's nearest parent container
+ * (group or array). Returns `undefined` when the field is at the form root
+ * with no parent container.
+ *
+ * @internal
+ */
+export function buildNearestGroupPath(context: CollectionContext): string | undefined {
+  if (context.arrayPath && context.groupPath) {
+    return `${context.arrayPath}.$.${context.groupPath}`;
+  }
+  return context.arrayPath ?? context.groupPath;
+}
+
+/**
+ * Substitutes `$self` / `$self.X` / `$group` / `$group.X` tokens in a
+ * `dependsOn` array with their resolved absolute paths. Non-token strings
+ * pass through unchanged.
+ *
+ * Throws `DynamicFormError` when `$group` is used on a field at the form
+ * root (no parent container exists).
+ *
+ * @internal
+ */
+export function resolveTokenDependencies(deps: string[], effectiveFieldKey: string, context: CollectionContext): string[] {
+  return deps.map((dep) => {
+    if (dep === SELF_DEPENDENCY_TOKEN) {
+      return effectiveFieldKey;
+    }
+    if (dep.startsWith(`${SELF_DEPENDENCY_TOKEN}.`)) {
+      return effectiveFieldKey + dep.slice(SELF_DEPENDENCY_TOKEN.length);
+    }
+    if (dep === GROUP_DEPENDENCY_TOKEN || dep.startsWith(`${GROUP_DEPENDENCY_TOKEN}.`)) {
+      const groupPath = buildNearestGroupPath(context);
+      if (!groupPath) {
+        throw new DynamicFormError(
+          `Derivation for '${effectiveFieldKey}' uses '${dep}' but the field has no parent group or array. ` +
+            `'${GROUP_DEPENDENCY_TOKEN}' is only valid for fields nested inside a group or array container.`,
+        );
+      }
+      if (dep === GROUP_DEPENDENCY_TOKEN) return groupPath;
+      return groupPath + dep.slice(GROUP_DEPENDENCY_TOKEN.length);
+    }
+    return dep;
+  });
+}
+
+/**
+ * Standard `traverseFieldsWithContext` step handlers for derivation
+ * collection. Array boundary resets `groupPath`; group descent appends the
+ * field's key; layout containers leave context unchanged.
+ *
+ * @internal
+ */
+export const CONTEXT_TRAVERSAL_OPTIONS = {
+  onArrayChild: (_parent: CollectionContext, field: { key: string }) => ({ arrayPath: field.key, groupPath: undefined }),
+  onGroupChild: (parent: CollectionContext, field: { key?: string }) => {
+    if (!field.key) return {};
+    return { groupPath: parent.groupPath ? `${parent.groupPath}.${field.key}` : field.key };
+  },
+};
